Move MainRoute to common module
