+++ app/src/main/java/com/epmedu/animeal/MainActivity.kt
@@ -7,10 +7,10 @@ import androidx.compose.animation.ExperimentalAnimationApi
 import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.ui.Modifier
 import androidx.core.view.WindowCompat
+import com.epmedu.animeal.common.route.MainRoute
 import com.epmedu.animeal.foundation.theme.AnimealTheme
 import com.epmedu.animeal.foundation.theme.TransparentSystemUi
 import com.epmedu.animeal.navigation.AnimatedScreenNavHost
-import com.epmedu.animeal.navigation.route.MainRoute
 import com.epmedu.animeal.signup.SignUpFlow
 import com.epmedu.animeal.splash.presentation.SplashScreen
 import dagger.hilt.android.AndroidEntryPoint
