+++ feature/mainflow/moreflow/account/src/main/java/com/epmedu/animeal/more/account/AccountScreen.kt
@@ -3,12 +3,12 @@ package com.epmedu.animeal.more.account
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.hilt.navigation.compose.hiltViewModel
+import com.epmedu.animeal.common.route.MainRoute
 import com.epmedu.animeal.extensions.currentOrThrow
 import com.epmedu.animeal.foundation.bottombar.BottomBarVisibility
 import com.epmedu.animeal.foundation.bottombar.BottomBarVisibilityState.HIDDEN
 import com.epmedu.animeal.more.account.viewmodel.AccountViewModel
 import com.epmedu.animeal.navigation.navigator.LocalNavigator
-import com.epmedu.animeal.navigation.route.MainRoute
 
 @Composable
 fun AccountScreen() {
